Fixed a graphics error, makefile improvment

+++ FLTK/graphic/Frame.cpp
@@ -42,7 +42,7 @@ void Frame::draw() {
 		glViewport(0, 0, pixel_w(), pixel_h());                      // Make our viewport the whole window
 		glMatrixMode(GL_PROJECTION);                           // Select The Projection Matrix
 		glLoadIdentity();                                      // Reset The Projection Matrix
-		gluPerspective(45.0f, pixel_w() / pixel_h(), 1, 150.0f);
+		gluPerspective(45.0f, ((float)pixel_w()) / pixel_h(), 1, 150.0f);
 		glMatrixMode(GL_MODELVIEW);                            // Select The Modelview Matrix
 		glLoadIdentity();                                      // Reset The Modelview Matrix
 		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);    // Clear The Screen And The Depth Buffer
